Portfolio/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-environ==0.4.5
django-ipware==2.1.0
django-request-logging==0.7.0
djangorestframework==3.10.1
idna==2.8
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
